+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `--no-color` option to disable color output [#141](https://github.com/Senth/minecraft-mod-manager/issues/141)
+- Runtime cache to save on requests to Curse/Modrinth APIs
 
 ### Changed